+++ tests/MyStringScalar.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Tests;
+namespace MLL\GraphQLScalars\Tests;
 
 use MLL\GraphQLScalars\StringScalar;
 
-class MyStringScalar extends StringScalar
+final class MyStringScalar extends StringScalar
 {
     public ?string $description = 'Bar';
 
     protected function isValid(string $stringValue): bool
     {
-        return $stringValue === 'foo';
+        return 'foo' === $stringValue;
     }
 }